created abstract class for different verbose serializers
